+++ sandbox/audit-ms-example/src/controllers/index.ts
@@ -2,4 +2,5 @@
 //
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
+export * from './ping.controller';
 export * from './to-do.controller';