+++ sandbox/auth-public-private-client/migrations/auth/migrations/20240502140840-seed-up.js
@@ -19,11 +19,7 @@ exports.setup = function (options, seedLink) {
 };
 
 exports.up = function (db) {
-  var filePath = path.join(
-    __dirname,
-    'sqls',
-    '20240502140840-seed-up.sql',
-  );
+  var filePath = path.join(__dirname, 'sqls', '20240502140840-seed-up.sql');
   return new Promise(function (resolve, reject) {
     fs.readFile(filePath, {encoding: 'utf-8'}, function (err, data) {
       if (err) return reject(err);
@@ -37,11 +33,7 @@ exports.up = function (db) {
 };
 
 exports.down = function (db) {
-  var filePath = path.join(
-    __dirname,
-    'sqls',
-    '20240502140840-seed-up.sql',
-  );
+  var filePath = path.join(__dirname, 'sqls', '20240502140840-seed-up.sql');
   return new Promise(function (resolve, reject) {
     fs.readFile(filePath, {encoding: 'utf-8'}, function (err, data) {
       if (err) return reject(err);