+++ sandbox/auth-public-private-client/src/datasources/auth.datasource.ts
@@ -1,6 +1,6 @@
-import { inject, lifeCycleObserver, LifeCycleObserver } from '@loopback/core';
-import { juggler } from '@loopback/repository';
-import { AuthDbSourceName } from '@sourceloop/authentication-service';
+import {inject, lifeCycleObserver, LifeCycleObserver} from '@loopback/core';
+import {juggler} from '@loopback/repository';
+import {AuthDbSourceName} from '@sourceloop/authentication-service';
 
 const DEFAULT_MAX_CONNECTIONS = 25;
 const DEFAULT_DB_IDLE_TIMEOUT_MILLIS = 60000;
@@ -14,7 +14,7 @@ const config = {
   user: process.env.DB_USER,
   schema: process.env.DB_SCHEMA,
   password: process.env.DB_PASSWORD,
-  database: process.env.DB_DATABASE
+  database: process.env.DB_DATABASE,
 };
 
 // Observe application's life cycle to disconnect the datasource when
@@ -22,16 +22,16 @@ const config = {
 // gracefully. The `stop()` method is inherited from `juggler.DataSource`.
 // Learn more at https://loopback.io/doc/en/lb4/Life-cycle.html
 @lifeCycleObserver('datasource')
-export class AuthDataSource extends juggler.DataSource
-  implements LifeCycleObserver {
-
+export class AuthDataSource
+  extends juggler.DataSource
+  implements LifeCycleObserver
+{
   static readonly dataSourceName = AuthDbSourceName;
-
 
   static readonly defaultConfig = config;
 
   constructor(
-    @inject('datasources.config.auth', { optional: true })
+    @inject('datasources.config.auth', {optional: true})
     dsConfig: object = config,
   ) {
     if (!!+(process.env.ENABLE_DB_CONNECTION_POOLING ?? 0)) {
@@ -46,7 +46,7 @@ export class AuthDataSource extends juggler.DataSource
         ),
       };
 
-      dsConfig = { ...dsConfig, ...dbPool };
+      dsConfig = {...dsConfig, ...dbPool};
     }
 
     super(dsConfig);