+++ sandbox/feature-toggle-example/src/controllers/index.ts
@@ -3,3 +3,4 @@
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 export * from './feature-toggle.controller';
+export * from './ping.controller';