+++ services/feature-toggle-service/migrations/pg/migrations/sqls/20230612141912-update-tables-up.sql
@@ -1,7 +1,7 @@
 /* Replace with your SQL commands */
 
 ALTER TABLE main.features
-ADD metadata text,  
+ADD metadata TEXT,  
 ADD created_by varchar(100),
 ADD modified_by varchar(100),
 ADD created_on           timestamptz DEFAULT CURRENT_TIMESTAMP NOT NULL ,
@@ -19,7 +19,7 @@ ADD deleted              bool DEFAULT false NOT NULL ,
 ADD deleted_on           timestamptz   ,
 ADD deleted_by           uuid   ;
 
-ALTER TABLE main.feature_toggles
+ALTER TABLE main.feature_values
 ADD created_by varchar(100),
 ADD modified_by varchar(100),
 ADD created_on           timestamptz DEFAULT CURRENT_TIMESTAMP NOT NULL ,