+++ services/feature-toggle-service/src/controllers/feature-values.controller.ts
@@ -11,13 +11,13 @@ import {
   Where,
 } from '@loopback/repository';
 import {
-  post,
-  param,
+  del,
   get,
   getModelSchemaRef,
+  param,
   patch,
+  post,
   put,
-  del,
   requestBody,
 } from '@loopback/rest';
 import {
@@ -28,30 +28,30 @@ import {
 import {authenticate, STRATEGY} from 'loopback4-authentication';
 import {authorize} from 'loopback4-authorization';
 import {PermissionKey} from '../enums';
-import {FeatureToggle} from '../models';
-import {FeatureToggleRepository} from '../repositories';
+import {FeatureValues} from '../models';
+import {FeatureValuesRepository} from '../repositories';
 
-const basePath = '/featuretoggles';
-export class FeatureToggleController {
+const basePath = '/feature-values';
+export class FeatureValuesController {
   constructor(
-    @repository(FeatureToggleRepository)
-    public featureToggleRepository: FeatureToggleRepository,
+    @repository(FeatureValuesRepository)
+    public featureValuesRepository: FeatureValuesRepository,
   ) {}
 
   @authenticate(STRATEGY.BEARER)
   @authorize({
     permissions: [
-      PermissionKey.CreateFeatureToggle,
-      PermissionKey.CreateFeatureToggleNum,
+      PermissionKey.CreateFeatureValues,
+      PermissionKey.CreateFeatureValuesNum,
     ],
   })
   @post(basePath, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
       [STATUS_CODE.OK]: {
-        description: 'FeatureToggle model instance',
+        description: 'FeatureValues model instance',
         content: {
-          [CONTENT_TYPE.JSON]: {schema: getModelSchemaRef(FeatureToggle)},
+          [CONTENT_TYPE.JSON]: {schema: getModelSchemaRef(FeatureValues)},
         },
       },
     },
@@ -60,56 +60,56 @@ export class FeatureToggleController {
     @requestBody({
       content: {
         [CONTENT_TYPE.JSON]: {
-          schema: getModelSchemaRef(FeatureToggle, {
-            title: 'NewFeatureToggle',
+          schema: getModelSchemaRef(FeatureValues, {
+            title: 'NewFeatureValues',
           }),
         },
       },
     })
-    featureToggle: FeatureToggle,
-  ): Promise<FeatureToggle> {
-    return this.featureToggleRepository.create(featureToggle);
+    featureValues: FeatureValues,
+  ): Promise<FeatureValues> {
+    return this.featureValuesRepository.create(featureValues);
   }
 
   @authenticate(STRATEGY.BEARER)
   @authorize({
     permissions: [
-      PermissionKey.ViewFeatureToggle,
-      PermissionKey.ViewFeatureToggleNum,
+      PermissionKey.ViewFeatureValues,
+      PermissionKey.ViewFeatureValuesNum,
     ],
   })
   @get(`${basePath}/count`, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
       [STATUS_CODE.OK]: {
-        description: 'FeatureToggle model count',
+        description: 'FeatureValues model count',
         content: {[CONTENT_TYPE.JSON]: {schema: CountSchema}},
       },
     },
   })
   async count(
-    @param.where(FeatureToggle) where?: Where<FeatureToggle>,
+    @param.where(FeatureValues) where?: Where<FeatureValues>,
   ): Promise<Count> {
-    return this.featureToggleRepository.count(where);
+    return this.featureValuesRepository.count(where);
   }
 
   @authenticate(STRATEGY.BEARER)
   @authorize({
     permissions: [
-      PermissionKey.ViewFeatureToggle,
-      PermissionKey.ViewFeatureToggleNum,
+      PermissionKey.ViewFeatureValues,
+      PermissionKey.ViewFeatureValuesNum,
     ],
   })
   @get(basePath, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
       [STATUS_CODE.OK]: {
-        description: 'Array of FeatureToggle model instances',
+        description: 'Array of FeatureValues model instances',
         content: {
           [CONTENT_TYPE.JSON]: {
             schema: {
               type: 'array',
-              items: getModelSchemaRef(FeatureToggle, {includeRelations: true}),
+              items: getModelSchemaRef(FeatureValues, {includeRelations: true}),
             },
           },
         },
@@ -117,23 +117,23 @@ export class FeatureToggleController {
     },
   })
   async find(
-    @param.filter(FeatureToggle) filter?: Filter<FeatureToggle>,
-  ): Promise<FeatureToggle[]> {
-    return this.featureToggleRepository.find(filter);
+    @param.filter(FeatureValues) filter?: Filter<FeatureValues>,
+  ): Promise<FeatureValues[]> {
+    return this.featureValuesRepository.find(filter);
   }
 
   @authenticate(STRATEGY.BEARER)
   @authorize({
     permissions: [
-      PermissionKey.UpdateFeatureToggle,
-      PermissionKey.UpdateFeatureToggleNum,
+      PermissionKey.UpdateFeatureValues,
+      PermissionKey.UpdateFeatureValuesNum,
     ],
   })
   @patch(basePath, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
       [STATUS_CODE.OK]: {
-        description: 'FeatureToggle PATCH success count',
+        description: 'FeatureValues PATCH success count',
         content: {[CONTENT_TYPE.JSON]: {schema: CountSchema}},
       },
     },
@@ -142,31 +142,31 @@ export class FeatureToggleController {
     @requestBody({
       content: {
         [CONTENT_TYPE.JSON]: {
-          schema: getModelSchemaRef(FeatureToggle, {partial: true}),
+          schema: getModelSchemaRef(FeatureValues, {partial: true}),
         },
       },
     })
-    featureToggle: FeatureToggle,
-    @param.where(FeatureToggle) where?: Where<FeatureToggle>,
+    featureValues: FeatureValues,
+    @param.where(FeatureValues) where?: Where<FeatureValues>,
   ): Promise<Count> {
-    return this.featureToggleRepository.updateAll(featureToggle, where);
+    return this.featureValuesRepository.updateAll(featureValues, where);
   }
 
   @authenticate(STRATEGY.BEARER)
   @authorize({
     permissions: [
-      PermissionKey.ViewFeatureToggle,
-      PermissionKey.ViewFeatureToggleNum,
+      PermissionKey.ViewFeatureValues,
+      PermissionKey.ViewFeatureValuesNum,
     ],
   })
   @get(`${basePath}/{id}`, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
       [STATUS_CODE.OK]: {
-        description: 'FeatureToggle model instance',
+        description: 'FeatureValues model instance',
         content: {
           [CONTENT_TYPE.JSON]: {
-            schema: getModelSchemaRef(FeatureToggle, {includeRelations: true}),
+            schema: getModelSchemaRef(FeatureValues, {includeRelations: true}),
           },
         },
       },
@@ -174,24 +174,24 @@ export class FeatureToggleController {
   })
   async findById(
     @param.path.string('id') id: string,
-    @param.filter(FeatureToggle, {exclude: 'where'})
-    filter?: FilterExcludingWhere<FeatureToggle>,
-  ): Promise<FeatureToggle> {
-    return this.featureToggleRepository.findById(id, filter);
+    @param.filter(FeatureValues, {exclude: 'where'})
+    filter?: FilterExcludingWhere<FeatureValues>,
+  ): Promise<FeatureValues> {
+    return this.featureValuesRepository.findById(id, filter);
   }
 
   @authenticate(STRATEGY.BEARER)
   @authorize({
     permissions: [
-      PermissionKey.UpdateFeatureToggle,
-      PermissionKey.UpdateFeatureToggleNum,
+      PermissionKey.UpdateFeatureValues,
+      PermissionKey.UpdateFeatureValuesNum,
     ],
   })
   @patch(`${basePath}/{id}`, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
       [STATUS_CODE.NO_CONTENT]: {
-        description: 'FeatureToggle PATCH success',
+        description: 'FeatureValues PATCH success',
       },
     },
   })
@@ -200,53 +200,53 @@ export class FeatureToggleController {
     @requestBody({
       content: {
         [CONTENT_TYPE.JSON]: {
-          schema: getModelSchemaRef(FeatureToggle, {partial: true}),
+          schema: getModelSchemaRef(FeatureValues, {partial: true}),
         },
       },
     })
-    featureToggle: FeatureToggle,
+    featureValues: FeatureValues,
   ): Promise<void> {
-    await this.featureToggleRepository.updateById(id, featureToggle);
+    await this.featureValuesRepository.updateById(id, featureValues);
   }
 
   @authenticate(STRATEGY.BEARER)
   @authorize({
     permissions: [
-      PermissionKey.UpdateFeatureToggle,
-      PermissionKey.UpdateFeatureToggleNum,
+      PermissionKey.UpdateFeatureValues,
+      PermissionKey.UpdateFeatureValuesNum,
     ],
   })
   @put(`${basePath}/{id}`, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
       [STATUS_CODE.NO_CONTENT]: {
-        description: 'FeatureToggle PUT success',
+        description: 'FeatureValues PUT success',
       },
     },
   })
   async replaceById(
     @param.path.string('id') id: string,
-    @requestBody() featureToggle: FeatureToggle,
+    @requestBody() featureValues: FeatureValues,
   ): Promise<void> {
-    await this.featureToggleRepository.replaceById(id, featureToggle);
+    await this.featureValuesRepository.replaceById(id, featureValues);
   }
 
   @authenticate(STRATEGY.BEARER)
   @authorize({
     permissions: [
-      PermissionKey.DeleteFeatureToggle,
-      PermissionKey.DeleteFeatureToggleNum,
+      PermissionKey.DeleteFeatureValues,
+      PermissionKey.DeleteFeatureValuesNum,
     ],
   })
   @del(`${basePath}/{id}`, {
     security: OPERATION_SECURITY_SPEC,
     responses: {
       [STATUS_CODE.NO_CONTENT]: {
-        description: 'FeatureToggle DELETE success',
+        description: 'FeatureValues DELETE success',
       },
     },
   })
   async deleteById(@param.path.string('id') id: string): Promise<void> {
-    await this.featureToggleRepository.deleteById(id);
+    await this.featureValuesRepository.deleteById(id);
   }
 }