+++ services/feature-toggle-service/src/controllers/index.ts
@@ -2,6 +2,6 @@
 //
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
+export * from './feature-values.controller';
 export * from './feature.controller';
 export * from './strategy.controller';
-export * from './feature-toggle.controller';