+++ sandbox/chat-notification-pubnub-example/facade/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.5.0 (2024-08-05)
+
+* chore(deps): version bump commit (#2111) ([c87bc75](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/c87bc75)), closes [#2111](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/2111) [#00](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/00)
+* chore(sandbox): fix the docker compose for sandbox (#2132) ([13a4945](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/13a4945)), closes [#2132](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/2132) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962)
+
+
+
+
+
 ## <small>0.4.10 (2024-04-05)</small>
 
 **Note:** Version bump only for package @sourceloop/chat-notif-example_facade