+++ services/chat-service/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 12.0.0 (2024-08-05)
+
+* chore(all-services): ping api in sandbox examples (#2134) ([b69cda1](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/b69cda1)), closes [#2134](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/2134) [#2133](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/2133) [#2133](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/2133)
+* chore(sandbox): fix the docker compose for sandbox (#2132) ([13a4945](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/13a4945)), closes [#2132](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/2132) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962) [#1962](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/1962)
+
+
+
+
+
 ## <small>11.0.1 (2024-06-11)</small>
 
 * chore(deps): loopback and sourceloop version update (#2108) ([9cd8962](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/9cd8962)), closes [#2108](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/2108) [#2104](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/2104) [#2104](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/2104)